Move parsing to Inventory::Parser
